+++ TaskApp/Assets/Scripts/BonusesLojic/Weapons/Guns/IGun.cs
@@ -1,11 +1,12 @@
 ﻿using System;
+using UnityEngine;
 
 namespace BonusLogic.Weapon
 {
     public interface IGun
     {
         string Name { get; }
-        float Damage { get; }
+        int Damage { get; }
         float Speed { get; }
         float Range { get; }
         void Shoot();
